[OPIK-425] Add images to dataset/experiment items sidebar
